+++ core/src/main/java/org/realityforge/proton/SuppressWarningsUtil.java
@@ -15,10 +15,8 @@ import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.annotation.processing.ProcessingEnvironment;
-import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ElementKind;
-import javax.lang.model.element.TypeElement;
 import javax.lang.model.type.ArrayType;
 import javax.lang.model.type.DeclaredType;
 import javax.lang.model.type.ExecutableType;
@@ -191,7 +189,7 @@ public final class SuppressWarningsUtil
     // short cut traversing types by checking whether additionalSuppressions match
     final boolean hasRawTypes =
       additionalSuppressions.contains( "rawtypes" ) ||
-      types.stream().anyMatch( t -> hasRawTypes( processingEnv, t ) );
+      types.stream().anyMatch( t -> TypesUtil.hasRawTypes( processingEnv, t ) );
 
     final boolean hasDeprecatedTypes =
       additionalSuppressions.contains( "deprecation" ) ||
@@ -283,57 +281,6 @@ public final class SuppressWarningsUtil
               ElementKind.PACKAGE != element.getEnclosingElement().getKind() )
     {
       return isElementDeprecated( element.getEnclosingElement() );
-    }
-    else
-    {
-      return false;
-    }
-  }
-
-  private static boolean hasRawTypes( @Nonnull final ProcessingEnvironment processingEnv,
-                                      @Nonnull final TypeMirror type )
-  {
-    final TypeKind kind = type.getKind();
-    if ( TypeKind.TYPEVAR == kind )
-    {
-      final TypeVariable typeVariable = (TypeVariable) type;
-      return hasRawTypes( processingEnv, typeVariable.getLowerBound() ) ||
-             hasRawTypes( processingEnv, typeVariable.getUpperBound() );
-    }
-    else if ( TypeKind.ARRAY == kind )
-    {
-      return hasRawTypes( processingEnv, ( (ArrayType) type ).getComponentType() );
-    }
-    else if ( TypeKind.DECLARED == kind )
-    {
-      final DeclaredType declaredType = (DeclaredType) type;
-      final int typeArgumentCount = declaredType.getTypeArguments().size();
-      final TypeElement typeElement = (TypeElement) processingEnv.getTypeUtils().asElement( type );
-      if ( typeArgumentCount != typeElement.getTypeParameters().size() )
-      {
-        return true;
-      }
-      else
-      {
-        return declaredType
-          .getTypeArguments()
-          .stream()
-          .anyMatch( t -> hasRawTypes( processingEnv, t ) );
-      }
-    }
-    else if ( TypeKind.EXECUTABLE == kind )
-    {
-      final ExecutableType executableType = (ExecutableType) type;
-      return hasRawTypes( processingEnv, executableType.getReturnType() ) ||
-             executableType.getTypeVariables()
-               .stream()
-               .anyMatch( t -> hasRawTypes( processingEnv, t ) ) ||
-             executableType.getThrownTypes()
-               .stream()
-               .anyMatch( t -> hasRawTypes( processingEnv, t ) ) ||
-             executableType.getParameterTypes()
-               .stream()
-               .anyMatch( t -> hasRawTypes( processingEnv, t ) );
     }
     else
     {