+++ core/src/main/java/org/realityforge/proton/TypesUtil.java
@@ -2,10 +2,14 @@ package org.realityforge.proton;
 
 import java.util.List;
 import javax.annotation.Nonnull;
+import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.TypeElement;
+import javax.lang.model.type.ArrayType;
 import javax.lang.model.type.DeclaredType;
+import javax.lang.model.type.ExecutableType;
 import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
+import javax.lang.model.type.TypeVariable;
 
 public final class TypesUtil
 {
@@ -78,6 +82,57 @@ public final class TypesUtil
         }
       }
       return false;
+    }
+    else
+    {
+      return false;
+    }
+  }
+
+  public static boolean hasRawTypes( @Nonnull final ProcessingEnvironment processingEnv,
+                                     @Nonnull final TypeMirror type )
+  {
+    final TypeKind kind = type.getKind();
+    if ( TypeKind.TYPEVAR == kind )
+    {
+      final TypeVariable typeVariable = (TypeVariable) type;
+      return hasRawTypes( processingEnv, typeVariable.getLowerBound() ) ||
+             hasRawTypes( processingEnv, typeVariable.getUpperBound() );
+    }
+    else if ( TypeKind.ARRAY == kind )
+    {
+      return hasRawTypes( processingEnv, ( (ArrayType) type ).getComponentType() );
+    }
+    else if ( TypeKind.DECLARED == kind )
+    {
+      final DeclaredType declaredType = (DeclaredType) type;
+      final int typeArgumentCount = declaredType.getTypeArguments().size();
+      final TypeElement typeElement = (TypeElement) processingEnv.getTypeUtils().asElement( type );
+      if ( typeArgumentCount != typeElement.getTypeParameters().size() )
+      {
+        return true;
+      }
+      else
+      {
+        return declaredType
+          .getTypeArguments()
+          .stream()
+          .anyMatch( t -> hasRawTypes( processingEnv, t ) );
+      }
+    }
+    else if ( TypeKind.EXECUTABLE == kind )
+    {
+      final ExecutableType executableType = (ExecutableType) type;
+      return hasRawTypes( processingEnv, executableType.getReturnType() ) ||
+             executableType.getTypeVariables()
+               .stream()
+               .anyMatch( t -> hasRawTypes( processingEnv, t ) ) ||
+             executableType.getThrownTypes()
+               .stream()
+               .anyMatch( t -> hasRawTypes( processingEnv, t ) ) ||
+             executableType.getParameterTypes()
+               .stream()
+               .anyMatch( t -> hasRawTypes( processingEnv, t ) );
     }
     else
     {