+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Expose `TypesUtil.hasRawTypes()` helper method.
+
 ### [v0.48](https://github.com/realityforge/proton/tree/v0.48) (2020-05-19) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.47...v0.48)
 
 Changes in this release: